requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
groq==0.9.0
neo4j==5.20.0
redis==5.0.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.26.4
pydantic==2.7.1
